+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Nothing so far
 * `Document#expiration_at` for OFMX
 * Generic features as raw XML (e.g. extracted from another AIXM/OFMX file)
 * Convenience combinator `AIXM::Schedule::DateTime`
+* Shortcuts `AIXM::BEGINNING_OF_DAY` and `AIXM::END_OF_DAY`
 
 ## 1.2.1